+++ src/index.ts
@@ -9,5 +9,8 @@ export * from './services';
 export * from './router';
 export * from './server';
 
+import { Request, Response } from 'express';
+export { Request, Response };
+
 import { Miter } from './server/miter';
 export default Miter;