+++ src/services/orm.service.ts
@@ -1,16 +1,24 @@
 import { Injector } from '../core/injector';
 import { Service } from '../decorators/services/service.decorator';
+import { Logger } from './logger';
 import { ClsNamespaceService } from './cls-namespace.service';
 import { TransactionT } from '../core/transaction';
 
-@Service()
+@Service({
+    provide: {
+        useCallback: (logger: Logger) => {
+            logger.error(`Failed to inject ORM service. There is no default implementation. Did you look into miter-sequelize?`);
+            return null;
+        },
+        deps: [Logger],
+        cache: true
+    }
+})
 export class ORMService {
     constructor(
         private _injector: Injector,
         private _namespace: ClsNamespaceService
-    ) {
-        if (this.constructor === ORMService) throw new Error(`Failed to start ORM service. There is no default implementation. Did you look into miter-sequelize?`);
-    }
+    ) { }
     
     protected get injector(): Injector {
         return this._injector;